requirements.txt
asgiref == 3.2.3
Django == 3.0.2
djangorestframework == 3.11.0
Pillow == 7.0.0
psycopg2 == 2.8.4
pytz == 2019.3
sqlparse == 0.3.0